--- Unholy/Assets/Scenes/BT/Ani Crtl.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d658e6df7f700e64fb12f0733f55aabc
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
