--- Unholy/Assets/Scenes/BT/Script.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0fff09f70cb071e49aa9cd6e36dbb953
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
